Data.BitMatrix

Description

A matrix of bits with some useful operations. Exposes row-based monomorphic maps, folds, and traversals. Intended to be used by multiple datatypes for space efficient character state encoding and packing.

Documentation

data BitMatrix

A data structure for storing a two dimensional array of bits. Exposes row based monomorphic mapping & folding.

It is important to note the endianness of BitMatrix. The bit at position (0,0) is displayed in the upper left hand corner when the BitMatrix is shown. The bit at position (i,x) will be of less significance than position (i+1,x), for the resulting xth BitVector row when calling rows on a BitMatrix. The bit at position (x,i) will be of less significance than position (x,i+1), for the resulting xth BitVector column when calling cols on a BitMatrix.

Construction

bitMatrix

Arguments

:: Word
Number of rows in the BitMatrix.

-> Word
Number of columns in the BitMatrix.

-> ((Word, Word) -> Bool)
Function to determine if a given index has a set bit.

-> BitMatrix

\( \mathcal{O} \left( m * n \right) \)

A generating function for a BitMatrix. Efficiently constructs a BitMatrix of the specified dimensions with each bit defined by the result of the supplied function.

Examples

>>> bitMatrix 3 2 $ const True
BitMatrix: 3 x 2
111
111
111

>>> bitMatrix 4 7 $ \(i,j) -> i == j || i + j == 6
BitMatrix: 4 x 7
1000001
0100010
0010100
0001000

>>> bitMatrix 5 10 $ \(i,j) -> odd i /= even j
BitMatrix: 5 x 10
1010101010
0101010101
1010101010
0101010101
1010101010

fromRows :: Foldable t => t BitVector -> BitMatrix

\( \mathcal{O} \left( m \right) \)

Construct a BitMatrix from a list of rows.

Querries

isSet :: BitMatrix -> (Word, Word) -> Bool

\( \mathcal{O} \left( 1 \right) \)

Test if a bit is set at the given indices.

isZeroMatrix :: BitMatrix -> Bool

\( \mathcal{O} \left( 1 \right) \)

Determines if there are no set bits in the BitMatrix

numCols :: BitMatrix -> Word

\( \mathcal{O} \left( 1 \right) \)

The number of columns in the BitMatrix

numRows :: BitMatrix -> Word

\( \mathcal{O} \left( 1 \right) \)

The number of rows in the BitMatrix

row :: BitMatrix -> Word -> BitVector

\( \mathcal{O} \left( 1 \right) \)

Retreives a single row of the BitMatrix. Allows for unsafe indexing.

rows :: BitMatrix -> [BitVector]

\( \mathcal{O} \left( m \right) \)

The rows of the BitMatrix

Specializations

expandRows :: BitMatrix -> BitVector

\( \mathcal{O} \left( 1 \right) \)

Extracts a bitvector with all cells concatenated in a row-major manner.

factorRows :: Word -> BitVector -> BitMatrix

\( \mathcal{O} \left( 1 \right) \)

Constructs a BitMatrix from a BitVector with all cells wrapped in a row-major manner.