Data.BitVector.LittleEndian

Description

A bit vector similar to BitVector from the bv package, however the Endianness reversed. This module defines Little-endian pseudo size-polymorphic bit-vectors.

Little-endian bit vectors are isomorphic to a [Bool] with the least significant bit at the head of the list and the most significant bit at the tail of the list.

Consequently, the endian-ness of a bit vector affects the Bits, FiniteBits, Semigroup, Monoid, MonoFoldable, and MonoTraversable instances.

If you want Bitvectors which are isomorphic to a [Bool] with the most significant bit at the head of the list and the least significant bit at the tail of the list, then you should use the bv package and not bv-little.

This module does not define numeric instances for BitVector. This is intentional! If you want to interact with a bit vector as an Integral value, convert the BitVector using either toSignedNumber or toUnsignedNumber.

Documentation

data BitVector

A little-endian bit-vector of non-negative dimension.

Construction from values

bitvector

Arguments

:: Integral v

=> Word
Bit vector dimension

-> v
Bit vector integral value, little-endian

-> BitVector

\( \mathcal{O} \left( 1 \right) \)

Create a bit-vector of non-negative dimension from an integral value. The integral value will be interpreted as little-endian so that the least significant bit of the integral value will be the value of the 0th index of the resulting bit-vector, and the most significant bit of the inegral value will be the largest index i such that bv testBit i is True.

Note that if the integral value exceeds the dimension, the set bits in the integral value that exceed the provided bit-vector dimension will be ignored.

>>> bitvector 7 96
[7]96

Queries

dimension :: BitVector -> Word

\( \mathcal{O} \left( 1 \right) \)

Get the dimension of a BitVector. Preferable over finiteBitSize as it returns a type which cannot represent a non-negative value.

>>> dimension [2]3
2

>>> dimension [4]12
4

isZeroVector :: BitVector -> Bool

\( \mathcal{O} \left( 1 \right) \)

Determine if _any_ bits are set in the BitVector. Preferable over '(0 ==) . popCount' as it is faster.

>>> isZeroVector [2]3
False

>>> isZeroVector [4]0
True

subRange :: (Word, Word) -> BitVector -> BitVector

\( \mathcal{O} \left( 1 \right) \)

Get the inclusive range of bits in BitVector as a new BitVector.

Numeric conversion

toSignedNumber :: Num a => BitVector -> a

\( \mathcal{O} \left( 1 \right) \)

2's complement value of a bit-vector.

>>> toSignedNumber [4]0
0

>>> toSignedNumber [4]3
3

>>> toSignedNumber [4]7
7

>>> toSignedNumber [4]8
-8

>>> toSignedNumber [4]12
-4

>>> toSignedNumber [4]15
-1

toUnsignedNumber :: Num a => BitVector -> a

\( \mathcal{O} \left( 1 \right) \)

Unsigned value of a bit-vector.

>>> toSignedNumber [4]0
0

>>> toSignedNumber [4]3
3

>>> toSignedNumber [4]7
7

>>> toSignedNumber [4]8
8

>>> toSignedNumber [4]12
12

>>> toSignedNumber [4]15
15

Bit stream conversion

fromBits :: Foldable f => f Bool -> BitVector

\( \mathcal{O} \left( n \right) \)

Create a bit-vector from a little-endian list of bits.

The following will hold:

length . takeWhile not === countLeadingZeros . fromBits

length . takeWhile not . reverse === countTrailingZeros . fromBits

>>> fromBits [True, False, False]
[3]1

toBits :: BitVector -> [Bool]

\( \mathcal{O} \left( n \right) \)

Create a little-endian list of bits from a bit-vector.

The following will hold:

length . takeWhile not . toBits === countLeadingZeros

length . takeWhile not . reverse . toBits === countTrailingZeros

>>> toBits [4]11
[True, True, False, True]