Data.Either.Custom

Description

Functions for collecting Left values of Either forming a Validation similar context.

Documentation

eitherValidation :: (Foldable t, Semigroup e) => t (Either e a) -> Either e [a]

\( \mathcal{O} \left( n \right) \)

Allows for the coalescence of Left values that form a Semigroup. Overrides short-circuiting behaviour of the standard monadic sequence definition. The validation context will evaluate the entire [Either e a] returning a Right value if and only if only Right values were present in the list of Eithers. If any Left values were present in the list of Eithers then a Left value is returned consisting of the an order preserving fold using the Semigroup operation (<>).

Examples

Basic usage:

>>> eitherValidation [Right 1, Right 2]
Right [1,2]

>>> eitherValidation [Right 1, Left "Hello", Right 2]
Left "Hello"

>>> eitherValidation [Right 1, Left "Hello", Right 2, Left "Good", Left "bye"]
Left "HelloGoodbye"

>>> eitherValidation [Left ("Love", "Hate"), Left (" you", " me")]
Left ("Love you", "Hate me")

eitherTValidation :: (Foldable t, Monad m, Semigroup e) => t (ExceptT e m a) -> ExceptT e m [a]

\( \mathcal{O} \left( n \right) \)

Works similarly to eitherValidation but within the MonadTrans context.