Data.List.Utility

Description

Functions for finding occurances of elements in a list.

Documentation

transpose :: (Applicative f, Applicative t, Semigroup (t a), Traversable t, Zip f) => t (f a) -> f (t a)

\( \mathcal{O} \left( n * k \right) \)

Takes two nested, linear-dimentional structures and transposes thier dimensions. It's like performing a matrix transpose operation, but more general.

_Example==

>>> transpose []
[]

>>> transpose [[1]]
[[1]]

>>> transpose [ [ 1, 2 ], [ 3, 4 ] ]
[[1,3],[2,4]]

>>> transpose [ [ 1, 2, 3 ], [ 4, 5, 6 ], [ 7, 8, 9] ]
[[1,4,7],[2,5,8],[3,6,9]]

>>> transpose [ [ 1, 2, 3, 0, 0 ], [ 4, 5, 6, 0 ], [ 7, 8, 9] ]
[[1,4,7],[2,5,8],[3,6,9]]

isSingleton :: Foldable t => t a -> Bool

\( \mathcal{O} \left( n * k \right) \) where \( k \) is the cost to convert the structure to a list in weak head normal form.

Determines whether a Foldable structure contains a single element.

_Example==

>>> isSingleton []
False

>>> isSingleton [ () ]
True

>>> isSingleton [ (), () ]
False

duplicates :: (Foldable t, Ord a) => t a -> [a]

\( \mathcal{O} \left( n * \log_2 n \right) \)

Returns the list of elements which are not unique in the input list.

_Example==

>>> duplicates "duplicate string"
"it"

>>> duplicates "GATACACATCAGATT"
"ACGT"

>>> duplicates [ 'A' .. 'Z' ]
[]

mostCommon :: (Foldable t, Ord a) => t a -> Maybe a

\( \mathcal{O} \left( n * \log_2 n \right) \)

Returns the element that occurs the most often in the list.

_Example==

>>> mostCommon "GATACACATCAGATT"
Just 'A'

>>> mostCommon "AABCDDDEFGGT"
Just 'D'

occurances :: (Foldable t, Ord a) => t a -> [(a, Int)]

\( \mathcal{O} \left( n * \log_2 n \right) \)

Returns a mapping of each unique element in the list paired with how often the element occurs in the list.

The elements are in descending order of occurance.

_Example==

>>> occurances "GATACACATCAGATT"
[('A',6),('T',4),('C',3),('G',2)]

>>> occurances "AABCDDDEFGGT"
[('D',3),('A',2),('G',2),('B',1),('C',1),('E',1),('F',1),('T',1)]

chunksOf :: Foldable t => Int -> t a -> [[a]]

\( \mathcal{O} \left( n * \right) \)

chunksOf is based on Text.chunksOf, but is more general.

_Example==

>>> chunksOf 3 [ 1 .. 13 ]
[[1,2,3],[4,5,6],[7,8,9],[10,11,12],[13]]

>>> chunksOf 5 [ 1 .. 13 ]
[[1,2,3,4,5],[6,7,8,9,10],[11,12,13]]

subsetOf :: (Foldable t, Foldable c, Ord a) => t a -> c a -> Bool

\( \mathcal{O} \left( n * \log_2 n \right) \)

Useful function to check subsets of lists.

_Example==

>>> [  5 .. 10 ] `subsetOf` [ 1 .. 13 ]
True

>>> [ 11 .. 15 ] `subsetOf` [ 1 .. 13 ]
False

equalityOf :: (Eq b, Foldable t) => (a -> b) -> t a -> Bool

\( \mathcal{O} \left( n \right) \)

Applies a transformation to each element of the structure and asserts that transformed values are equal for all elements of the structure.

See invariantTransformation if you need the equal value returned.

_Example==

>>> equalityOf (`mod` 10) [ 9, 19, 29, 39, 49 ]
True

>>> equalityOf (`mod`  7) [ 9, 19, 29, 39, 49 ]
False

invariantTransformation :: (Eq b, Foldable t) => (a -> b) -> t a -> Maybe b

\( \mathcal{O} \left( n \right) \)

Applies a transformation to each element of the structure. If every application of the transformation yields the same result value for each element of the structure then this function will return Just v where v is the invariant value across the transformation. If the transformation does not produce an invariant value accross the structure, or the structure is empty, this function returns Nothing.

See equalityOf if you want to discard the Just value.

_Example==

>>> invariantTransformation (`mod` 10) [ 9, 19, 29, 39, 49 ]
Just 9

>>> invariantTransformation (`mod`  7) [ 9, 19, 29, 39, 49 ]
Nothing

transitivePropertyHolds :: Foldable f => (a -> a -> Bool) -> f a -> Bool

\( \mathcal{O} \left( n \right) \)

Applies a transitive relation over a list and asserts that the relation holds for all pairs of elements. By applying the transitive property, we can assert that the relation hold in liner rather than quadratic time for the collection.

_Example==

>>> transitivePropertyHolds (\x y -> snd x >= fst y) [ (9,9), (8,7), (6,6), (6,5), (3,4), (3,0) ]
True 

pairwiseSequence :: (Foldable t, Foldable t') => (a -> a -> Bool) -> t (t' a) -> [[a]]

\( \mathcal{O} \left( m * n \right) \)

Provide a pairwise predicate used to filter elements and a nested structure. Returns the "product" of elements across the inner Foldable structure. Each result must have all elements satisfy the predicate when compared to all other elements.

_Example==

>>> pairwiseSequence (\x y -> snd x /= snd y) [[('A',1),('B',2)],[('X',1),('Y',2),('Z',3)],[('I',1),('J',2),('K',3),('L',4)]]
[[('A',1),('Y',2),('K',3)],[('A',1),('Y',2),('L',4)],[('A',1),('Z',3),('J',2)],[('A',1),('Z',3),('L',4)],[('B',2),('X',1),('K',3)],[('B',2),('X',1),('L',4)],[('B',2),('Z',3),('I',1)],[('B',2),('Z',3),('L',4)]]

maximaBy :: Foldable t => (a -> a -> Ordering) -> t a -> [a]

\( \mathcal{O} \left( n \right) \)

The largest elements of a possibly empty structure with respect to the given comparison function. If multiple elements are equal under the comparison function, then all the equally maximal elements are returned in the order they existed in the foldable structure.

The empty list is returned if and only if an empty structure was provided.

Similar to maximumBy. For non-empty structures, the following will hold:

head . maximaBy == maximumBy

minimaBy :: Foldable t => (a -> a -> Ordering) -> t a -> [a]

\( \mathcal{O} \left( n \right) \)

The smallest elements of a possibly empty structure with respect to the given comparison function. If multiple elements are equal under the comparison function, then all the equally minimal elements are returned in the order they existed in the foldable structure.

The empty list is returned if and only if an empty structure was provided.

Similar to minimumBy. For non-empty structures, the following will hold:

head . minimaBy == minimumBy