Data.MutualExclusionSet

Description

Set-like structure for a collection of elements where each included element implies the exclusion of another element.

The MutualExclusionSet holds a bijective mapping between a set of included elements and their corresponding mutually exclusive element. The construction allows for efficient access to included and excluded elements.

The following properties will always hold:

 invert . invert === id

 includedSet === excludedSet . invert

 excludedSet === includedSet . invert

 toList === toList . includedSet

 toList . invert === toList . excludedSet

 ∀ e, isIncluded e ==> not . isExcluded e

 ∀ e, isExcluded e ==> not . isIncluded e

 ∀ e, isIncluded e === isExcluded e . invert

 ∀ e, isExcluded e === isIncluded e . invert

 ∀ e, ∃! k, isIncluded e ==> excludedLookup k == Just e

 ∀ e, ∃! k, isExcluded e ==> includedLookup k == Just e

Documentation

data MutualExclusionSet a

Set of elements which are mutually exclusive.

Each element in the set implies that another element cannot be included in the set.

This is an efficient representation for construction and manipulation of a collection of mutually exclusive elements using an newtyped Bimap with some typeclass instance modifications.

Construction

singleton

Arguments

:: Eq a

=> a
Included element

-> a
Excluded element

-> MutualExclusionSet a

\( \mathcal{O} \left( 1 \right) \)

Construct a singleton MutualExclusionSet value by supplying an included element and the corresponding, mutually exclusive element.

Use the semigroup operator '(<>)' to merge singleton contexts into a larger MutualExclusionSet.

Deconstruct

excludedSet :: MutualExclusionSet a -> Set a

\( \mathcal{O} \left( n \right) \)

Retreive the list of excluded elements in the MutualExclusionSet.

includedSet :: MutualExclusionSet a -> Set a

\( \mathcal{O} \left( n \right) \)

Retreive the list of included elements in the MutualExclusionSet.

mutuallyExclusivePairs :: MutualExclusionSet a -> Set (a, a)

\( \mathcal{O} \left( n \right) \)

Retreive the list of mutually exclusive elements stored in the MutualExclusionSet.

The first element of the pair is the included element and the second element of pair is the excluded element.

Manipulation

invert :: MutualExclusionSet a -> MutualExclusionSet a

\( \mathcal{O} \left( 1 \right) \)

inverts the included and excluded elements of the MutualExclusionSet. The previously included elements become the excluded elements and the previously excluded elements become included elements.

This inversion function preserves the bijective relationship of the elements within MutualExclusionSet so that the following will always hold:

invert . invert === id

merge :: Ord a => MutualExclusionSet a -> MutualExclusionSet a -> MutualExclusionSet a

\( \mathcal{O} \left( m * \log_2 ( \frac {n}{m + 1} ) + n + m \right), m \leq n \)

Merge two mutual exclusion sets.

Perfoms an "union-like" operation.

prettyPrintMutualExclusionSet :: (Ord a, Show a) => MutualExclusionSet a -> String

Nicely render the MutualExclusionSet in a multi-line String.

Shows the internal state inluding bijectivity and mutual exclusivity violations.

Comparison / Queries

excludedLookup :: Ord a => a -> MutualExclusionSet a -> Maybe a

\( \mathcal{O} \left( \log_2 n \right) \)

Lookup an excluded key in the MutualExclusionSet.

If the provided element *is not* excluded, the result will be Nothing.

If the provided element *is* excluded, the result will be Just value, where value is corresponding included element.

includedLookup :: Ord a => a -> MutualExclusionSet a -> Maybe a

\( \mathcal{O} \left( \log_2 n \right) \)

Lookup an included key in the MutualExclusionSet.

If the provided element *is not* included, the result will be Nothing.

If the provided element *is* included, the result will be Just value, where value is corresponding excluded element.

isCoherent :: MutualExclusionSet a -> Bool

\( \mathcal{O} \left( 1 \right) \)

Determines if the MutualExclusionSet has a coherent construction.

There must be a bijective mapping between included and excluded elements.

No included element is also excluded

isExcluded :: Ord a => a -> MutualExclusionSet a -> Bool

\( \mathcal{O} \left( \log_2 n \right) \)

Query the MutualExclusionSet to determine if the provided element is excluded.

isIncluded :: Ord a => a -> MutualExclusionSet a -> Bool

\( \mathcal{O} \left( \log_2 n \right) \)

Query the MutualExclusionSet to determine if the provided element is included.

isPermissible :: Ord a => MutualExclusionSet a -> MutualExclusionSet a -> Bool

\( \mathcal{O} \left( m * \log_2 ( \frac {n}{m + 1} ) \right), m \leq n \)

Perform an operation to determine if a collection of elements is "permitted" by MutualExclusionSet, ie that the collection does not contain any elements which are excluded by the MutualExclusionSet.