Numeric.Extended

Description

Documentation

data ExtendedNatural

A type that extends the Natural numbers to include an infinity value.

This is a newtyped Word for efficiency purposes.

The following behavior should be noted:

ExtendedNatural has a domain of [0 .. (maxBound :: Word) - 1] <> infinity

unsafeToFinite (maxBound :: ExtendedNatural) == (maxBound :: Word) - 1

unsafeToFinite (minBound :: ExtendedNatural) == (minBound :: Word)

unsafeToFinite (infinity :: ExtendedNatural) == (maxBound :: Word)

ExtendedNatural does not throw an error on arithmetic underflows caused by subtracting a larger value from a smaller value. Instead, the result is floored at zero.

ExtendedNatural does not throw an error on arithmetic overflows caused by adding or multiplying two numbers that would produce a result larger than (maxBound :: ExtendedNatural). Instead, the result is capped at (maxBound :: ExtendedNatural).

An infinity value can only result from it's use as a constant, addition with infinity as an operand, multiplication with infinity as an operand subtraction with infinity as the minuend, or division with infinity as the denominator.

class ExtendedNumber n where

Conversion two and from a finite and infinite value domains.

The type-class definition holds an infinity value. Conversion from the extended type back to the finite type is inherently unsafe because you could attempt to coerce the infinity value to a finite value. It is the responsibility of the caller of unsafeToFinite to first check if the value is equal to infinity. This can be done with a value-level equality check '(==)'.

Minimal complete definition

unsafeToFinite, fromFinite, infinity

Methods

unsafeToFinite :: n -> Finite n

fromFinite :: Finite n -> n

infinity :: n

data ExtendedReal

A non-negative, real number extended to include infinity where:

infinity > maxBound

maxBound is defined as the largest representable finite value.

type family Finite (f :: *)

The finite type underlying the infinite extension type.